Refactor report service and remove unused code
