+++ src/app/core/models/platform/v1/reports-query-params.model.ts
@@ -1,0 +1,3 @@
+export interface ReportPlatformParams {
+  state?: string;
+}
