+++ src/app/core/models/report-api-params.model.ts
@@ -4,10 +4,6 @@ export interface ReportQueryParams {
   rp_id?: string;
 }
 
-export interface ReportPlatformParams {
-  state?: string;
-}
-
 export interface ReportApiParams {
   offset?: number;
   limit?: number;
