+++ src/app/core/services/report.service.ts
@@ -9,7 +9,7 @@ import { OrgSettings } from '../models/org-settings.model';
 import { PdfExport } from '../models/pdf-exports.model';
 import { PlatformReport } from '../models/platform/platform-report.model';
 import { ReportActions } from '../models/report-actions.model';
-import { ReportQueryParams, ReportPlatformParams } from '../models/report-api-params.model';
+import { ReportQueryParams } from '../models/report-api-params.model';
 import { ReportAutoSubmissionDetails } from '../models/report-auto-submission-details.model';
 import { ReportParams } from '../models/report-params.model';
 import { ReportPermission } from '../models/report-permission.model';
@@ -30,7 +30,6 @@ import { LaunchDarklyService } from './launch-darkly.service';
 import { NetworkService } from './network.service';
 import { PermissionsService } from './permissions.service';
 import { SpenderPlatformV1ApiService } from './spender-platform-v1-api.service';
-import { PlatformApiResponse } from '../models/platform/platform-api-response.model';
 import { StorageService } from './storage.service';
 import { TransactionService } from './transaction.service';
 import { UserEventService } from './user-event.service';
@@ -277,72 +276,6 @@ export class ReportService {
     return this.permissionsService
       .allowedActions('reports', ['approve', 'create', 'delete'], orgSettings)
       .pipe(catchError(() => [])) as Observable<Partial<ReportPermission>>;
-  }
-
-  @Cacheable()
-  getAllReportsByParams(queryParams: ReportPlatformParams = {}): Observable<PlatformReport[]> {
-    return of(queryParams).pipe(
-      // first we will get count of categories and on basis of that make parallel calls
-      mergeMap((queryParams) => this.getReportsCount(queryParams)),
-      mergeMap((count) => this.getReportsInParallel(count, queryParams))
-    );
-  }
-
-  getReportsByParams(
-    queryParams: ReportPlatformParams = {},
-    offset: number = 0,
-    limit: number = 1
-  ): Observable<PlatformApiResponse<PlatformReport>> {
-    const config = {
-      params: {
-        ...queryParams,
-        offset,
-        limit,
-      },
-    };
-    return this.spenderPlatformV1ApiService.get<PlatformApiResponse<PlatformReport>>('/reports', config);
-  }
-
-  getReportsCount(queryParams?: ReportPlatformParams): Observable<number> {
-    return this.getReportsByParams(queryParams, 0, 1).pipe(map((res) => res.count));
-  }
-
-  getReportsInParallel(
-    count: number,
-    queryParams: ReportPlatformParams,
-    offset: number = 0,
-    reports: PlatformReport[] = []
-  ): Observable<PlatformReport[]> {
-    const limit = 200;
-    const maxParallelCalls = 3;
-    if (count > limit) {
-      // Decides how many api calls should be sent at once in parallel, upto a max number
-      const numberOfCallsRequired = Math.ceil(count / limit);
-      const numberOfParallelCalls = Math.min(numberOfCallsRequired, maxParallelCalls);
-
-      return from(Array(numberOfParallelCalls).keys()).pipe(
-        mergeMap((i) => this.getReportsByParams(queryParams, offset + 200 * i, limit)),
-        map((response) => response.data),
-        // Since we receive [Response1, Response2, Response3], we will reduce them into array of reports.
-        reduce((accumulatedReports, currentReports) => [...accumulatedReports, ...currentReports]),
-        mergeMap((iterationCategories) => {
-          offset += 600;
-          reports = [...reports, ...iterationCategories];
-          const remainingCount = count - iterationCategories.length;
-          if (remainingCount > 0) {
-            // If we still have more data to fetch, call the same function again but with remaining count
-            return this.getReportsInParallel(remainingCount, queryParams, offset, reports);
-          } else {
-            return of(reports);
-          }
-        })
-      );
-    } else {
-      // This case will execute when the reports can be fetched in one api call only
-      return this.getReportsByParams(queryParams, offset, limit).pipe(
-        map((response) => [...reports, ...response.data])
-      );
-    }
   }
 
   getAutoSubmissionReportName(): Observable<string> {
